+++ test/viking/model/instance_properties/paramRoot.js
@@ -9,7 +9,7 @@
 
         Model = Viking.Model.extend('namespaced/model');
         model = new Model();
-        equal(model.paramRoot(), 'model');
+        equal(model.paramRoot(), 'namespaced_model');
     });
     
     test("instance.paramRoot returns underscored baseModel.modelName when used as STI", function() {